pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "taroni-cli"
version = "0.8.1" # 階層的設定ファイルの導入
authors = [
  { name="Taro.ai", email="sinzy0925@gmail.com" },
]
description = "A highly customizable and intelligent CLI assistant for the Google GenAI SDK"
requires-python = ">=3.9"
dependencies = [
    "google-genai",
    "python-dotenv",
    "tzdata",
    "httpx",
    "prompt-toolkit",
]

[project.scripts]
taroni = "taroni.cli:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["taroni"]